refactor(layout): split Layout.tsx into Header, Footer, UserMenu, DebugInfoPanel
Layout.tsx was a 444-line file mixing header nav, auth menu, footer, and a
debug overlay. Split into focused components (Layout is now 36 lines of
composition) per the Phase 4 roadmap target.

- Header.tsx owns the top ribbon, desktop nav, and mobile menu toggle
- UserMenu.tsx handles auth UI with a variant prop (desktop | mobile)
- Footer.tsx and DebugInfoPanel.tsx are pure static/view-only components
- Public Layout API and Storybook stories are unchanged
